refactor: flatten akademy-app directory structure
Move all modules to root level of akademy-app function for simplified imports and cleaner organization.

- Move services/* to root level
- Move routes/* to root level
- Move utils/* to root level
- Move middleware/* to root level
- Move schemas/* to root level
- Update import paths throughout codebase
