+++ functions/akademy-app/agreementChecks.ts
@@ -1,4 +1,4 @@
-import { SupabaseClient } from '@supabase/supabase-js';
+import { SupabaseClient } from 'jsr:@supabase/supabase-js@2';
 
 /**
  * Splits an array into batches of the specified size
