+++ functions/akademy-app/auth.ts
@@ -1,4 +1,4 @@
-import { createClient } from '@supabase/supabase-js';
+import { createClient } from 'jsr:@supabase/supabase-js@2';
 
 export async function getUserRoleLevel(token: string): Promise<number | null> {
   try {
