+++ functions/akademy-app/dataNormalization.ts
