+++ functions/akademy-app/helpers.ts
