+++ functions/akademy-app/index.ts
@@ -1,11 +1,11 @@
-import { Hono } from 'hono';
-import { cors } from 'hono/cors';
-import { HTTPException } from 'hono/http-exception';
+import { Hono } from 'jsr:@hono/hono@4';
+import { cors } from 'jsr:@hono/hono@4/cors';
+import { HTTPException } from 'jsr:@hono/hono@4/http-exception';
 import "jsr:@supabase/functions-js/edge-runtime.d.ts"
-import { strapiMigrationRoute } from "./routes/migration.ts";
-import { createUserFromAgreement } from "./routes/create-user.ts";
-import { resetUserPassword } from "./routes/reset-password.ts";
-import { deactivateUser } from "./routes/deactivate-user.ts";
+import { strapiMigrationRoute } from "./migration.ts";
+import { createUserFromAgreement } from "./create-user.ts";
+import { resetUserPassword } from "./reset-password.ts";
+import { deactivateUser } from "./deactivate-user.ts";
 import { 
 	searchUsers, 
 	sendNotification, 
@@ -16,8 +16,8 @@ import {
 	archiveNotification,
 	getNotificationPreferences,
 	updateNotificationPreferences 
-} from "./routes/notifications.ts";
-import { requireMinRoleLevel } from "./middleware/auth.ts";
+} from "./notifications.ts";
+import { requireMinRoleLevel } from "./middlewareAuth.ts";
 
 export const app = new Hono();
 
@@ -41,7 +41,6 @@ app.use('*', cors({
     maxAge: 86400,
 }));
 
-// Handle OPTIONS requests explicitly for all routes
 app.options('*', (c) => {
     return c.text('', 204);
 });
