+++ functions/akademy-app/logger.ts
